+++ src/renderer/src/vue-cal.d.ts
@@ -1,2 +1,0 @@
-declare module 'vue-cal'
-declare module 'vue-cal/dist/vuecal.css'